Core: Allow Build.DISPLAY or Build.FINGERPRINT to deterimine mIsUpgrade
Many CM devices use a static fingerprint that is only changed when a new stock
version is released, so every boot we see Android is starting instead of
upgrading.  Build.DISPLAY is updated each time the rom is built and will
signal framework that this is an upgrade even if Build.FINGERPRINT is unchanged.

Change-Id: I73c72bba6d3f6e678da3a2d05729ab8bc8eaa919