SystemUI: fix tiles per page when changing top tile type
When toggling whether to enlarge the first two tiles, the tiles were
retaining their previous pages and not properly moving to pages as
necessary (cauaing the first page to have more tiles than there should
be). Clearing out the tiles before to clear the states when the option
is toggled will lay the tiles out properly.

Ref: CYNGNOS-1644

Change-Id: I8f44faf4b9edf80418b24bea6f5906246be9ad07
Signed-off-by: Roman Birg <roman@cyngn.com>