SystemUI: fix contracting qs settings page
If notifications were present, isInQsArea would return true while
tapping in the notification area, blocking touches and not allowing the
user to grab the qs panel to drag it up.

Add a flag to not check for the notification area and use that for qs
settings interception, but leave the default
behavior for other uses.

Ref: CYNGNOS-1644

Change-Id: I092600c0ef9815e3a4fe2b8101c5b53ffa670275
Signed-off-by: Roman Birg <roman@cyngn.com>